Introduce CruiseControlUtil
